Refactor logs, shade-plugin name

+++ bukkit/src/main/java/com/github/theprogmatheus/craftlib/bukkit/PluginLibraryShader.java
@@ -1,6 +1,7 @@
 package com.github.theprogmatheus.craftlib.bukkit;
 
-import lombok.RequiredArgsConstructor;
+import com.github.theprogmatheus.craftlib.core.utils.FileUtils;
+import lombok.Getter;
 
 import java.io.*;
 import java.util.Collection;
@@ -11,17 +12,29 @@ import java.util.jar.Manifest;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
-@RequiredArgsConstructor
+@Getter
 public class PluginLibraryShader {
-
-    private final File shadeJar;
-    private final Collection<File> files;
-
 
     private static final String MAIN_CLASS_NAME = "com.github.theprogmatheus.craftlib.bukkit.shade.Main";
     private static final String PLUGIN_YML = "plugin.yml";
 
-    public File createShadedPluginJar() throws IOException {
+
+    private final File shadeJar;
+    private final Collection<File> files;
+    private final String shadeHash;
+
+
+    public PluginLibraryShader(File shadeJar, Collection<File> files) throws Exception {
+        this.shadeJar = shadeJar;
+        this.files = files;
+        this.shadeHash = FileUtils.hashFiles(files);
+    }
+
+
+    public File shade() throws Exception {
+        if (alreadyExists(true))
+            return this.shadeJar;
+
         try (JarOutputStream jarOutputStream = new JarOutputStream(new FileOutputStream(shadeJar), createManifest())) {
 
             for (File file : files) {
@@ -40,8 +53,21 @@ public class PluginLibraryShader {
             addDummyMain(jarOutputStream);
             addPluginYml(jarOutputStream);
         }
-
         return shadeJar;
+    }
+
+    public boolean alreadyExists(boolean checkHash) {
+        if (!checkHash)
+            return this.shadeJar.exists();
+
+        if (PluginFile.isValidJarFile(this.shadeJar)) {
+            PluginFile pluginFile = new PluginFile(this.shadeJar);
+            if (pluginFile.isValidPlugin()) {
+                String shadeHash = pluginFile.getPluginYaml().getString("shade-hash");
+                return (shadeHash != null && !shadeHash.isBlank()) && (this.shadeHash.equals(shadeHash));
+            }
+        }
+        return false;
     }
 
     private Manifest createManifest() {
@@ -52,7 +78,7 @@ public class PluginLibraryShader {
         return manifest;
     }
 
-    private void addDummyMain(JarOutputStream jos) throws IOException {
+    private void addDummyMain(JarOutputStream jos) throws Exception {
         String path = MAIN_CLASS_NAME.replace('.', '/') + ".class";
         try (InputStream is = getClass().getClassLoader().getResourceAsStream(path)) {
             if (is == null)
@@ -63,16 +89,17 @@ public class PluginLibraryShader {
         }
     }
 
-    private void addPluginYml(JarOutputStream jarOutputStream) throws IOException {
+    private void addPluginYml(JarOutputStream jarOutputStream) throws Exception {
         jarOutputStream.putNextEntry(new JarEntry(PLUGIN_YML));
         jarOutputStream.write(("""
                 load: STARTUP
-                depend: ["craftlib"]
-                name: craftlib-shade
+                depend: ["CraftLib"]
+                name: CraftLibs
                 main: %s
                 version: 1.0.0
                 api-version: 1.8
-                """.formatted(MAIN_CLASS_NAME)).getBytes());
+                shade-hash: %s
+                """.formatted(MAIN_CLASS_NAME, this.shadeHash)).getBytes());
         jarOutputStream.closeEntry();
     }
